refacto(3d): Move feature picking to another file
