+++ projects/hslayers-cesium/src/hscesium.service.ts
@@ -5,20 +5,14 @@ import BingMapsImageryProvider from 'cesium/Source/Scene/BingMapsImageryProvider
 import BingMapsStyle from 'cesium/Source/Scene/BingMapsStyle';
 import Camera from 'cesium/Source/Scene/Camera';
 import Cartesian3 from 'cesium/Source/Core/Cartesian3';
-import Cartographic from 'cesium/Source/Core/Cartographic';
 import CesiumTerrainProvider from 'cesium/Source/Core/CesiumTerrainProvider';
 import Ion from 'cesium/Source/Core/Ion';
-import Math from 'cesium/Source/Core/Math';
 import SceneMode from 'cesium/Source/Scene/SceneMode';
-import ScreenSpaceEventHandler from 'cesium/Source/Core/ScreenSpaceEventHandler';
-import ScreenSpaceEventType from 'cesium/Source/Core/ScreenSpaceEventType';
 import ShadowMode from 'cesium/Source/Scene/ShadowMode';
 import SkyBox from 'cesium/Source/Scene/SkyBox';
 import Viewer from 'cesium/Source/Widgets/Viewer/Viewer';
 import WebMercatorProjection from 'cesium/Source/Core/WebMercatorProjection';
 import createWorldTerrain from 'cesium/Source/Core/createWorldTerrain';
-import defined from 'cesium/Source/Core/defined';
-import when from 'cesium/Source/ThirdParty/when';
 import {HsCesiumCameraService} from './hscesium-camera.service';
 import {HsCesiumLayersService} from './hscesium-layers.service';
 import {HsCesiumTimeService} from './hscesium-time.service';
@@ -30,6 +24,7 @@ import {HsUtilsService} from 'hslayers-ng';
 import {Subject} from 'rxjs';
 
 import {HsCesiumConfig} from './hscesium-config.service';
+import {init as pickerInit} from './picking';
 
 @Injectable({
   providedIn: 'root',
@@ -206,67 +201,7 @@ export class HsCesiumService {
         });
       });
 
-      const handler = new ScreenSpaceEventHandler(this.viewer.scene.canvas);
-      handler.setInputAction((movement) => {
-        const pickRay = this.viewer.camera.getPickRay(movement.position);
-        const pickedObject = this.viewer.scene.pick(movement.position);
-        const featuresPromise =
-          this.viewer.imageryLayers.pickImageryLayerFeatures(
-            pickRay,
-            this.viewer.scene
-          );
-        if (pickedObject && pickedObject.id && pickedObject.id.onclick) {
-          pickedObject.id.onclick(pickedObject.id);
-          return;
-        }
-        if (!defined(featuresPromise)) {
-          if (console) {
-            console.log('No features picked.');
-          }
-        } else {
-          when(featuresPromise, (features) => {
-            let s = '';
-            if (features.length > 0) {
-              for (let i = 0; i < features.length; i++) {
-                s = s + features[i].data + '\n';
-              }
-            }
-            const iframe: any =
-              this.HsLayoutService.layoutElement.querySelector(
-                '.cesium-infoBox-iframe'
-              );
-            if (iframe) {
-              setTimeout(() => {
-                const innerDoc = iframe.contentDocument
-                  ? iframe.contentDocument
-                  : iframe.contentWindow.document;
-                innerDoc.querySelector(
-                  '.cesium-infoBox-description'
-                ).innerHTML = s.replace(/\n/gm, '<br/>');
-                iframe.style.height = 200 + 'px';
-              }, 1000);
-            }
-          });
-        }
-      }, ScreenSpaceEventType.LEFT_DOWN);
-
-      handler.setInputAction((movement) => {
-        const pickedObject = this.viewer.scene.pick(movement.position);
-        if (pickedObject && pickedObject.id && pickedObject.id.onmouseup) {
-          pickedObject.id.onmouseup(pickedObject.id);
-          return;
-        }
-      }, ScreenSpaceEventType.LEFT_UP);
-
-      handler.setInputAction(
-        rightClickLeftDoubleClick,
-        ScreenSpaceEventType.RIGHT_DOWN
-      );
-      handler.setInputAction(
-        rightClickLeftDoubleClick,
-        ScreenSpaceEventType.LEFT_DOUBLE_CLICK
-      );
-
+      pickerInit(this.viewer);
       this.HsEventBusService.cesiumLoads.next({viewer: viewer, service: this});
     } catch (ex) {
       console.error(ex);
@@ -353,29 +288,4 @@ export class HsCesiumService {
  */
 function loadSkyBoxSide(file) {
   return `${(<any>window).CESIUM_BASE_URL}Assets/Textures/SkyBox/${file}`;
-}
-
-/**
- * @param movement -
- */
-function rightClickLeftDoubleClick(movement) {
-  const pickRay = this.viewer.camera.getPickRay(movement.position);
-  const pickedObject = this.viewer.scene.pick(movement.position);
-
-  if (this.viewer.scene.pickPositionSupported) {
-    if (this.viewer.scene.mode === SceneMode.SCENE3D) {
-      const cartesian = this.viewer.scene.pickPosition(movement.position);
-      if (defined(cartesian)) {
-        const cartographic = Cartographic.fromCartesian(cartesian);
-        const longitudeString = Math.toDegrees(cartographic.longitude);
-        const latitudeString = Math.toDegrees(cartographic.latitude);
-        //TODO rewrite to subject
-        this.cesiumPositionClicked.next([longitudeString, latitudeString]);
-      }
-    }
-  }
-  if (pickedObject && pickedObject.id && pickedObject.id.onclick) {
-    pickedObject.id.onRightClick(pickedObject.id);
-    return;
-  }
 }
